requirements.txt
Django==4.1.5
django-crispy-forms==1.14.0
requests==2.27.1
Pillow==9.0.1
djangorestframework==3.13.1